requirements.txt
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
filelock==3.0.12
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
oauth2client==4.1.3
oauthlib==3.1.1
pip==21.1.1
platformdirs==2.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.2
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools==56.0.0
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
pymongo==3.5.1
